Initialise the repo.
Start out with a simple definition of what makes the accounts service
tick, mostly the Account type and the Storer interface. To make sure I'm
not crazy, I also implemented the Storer interface as an in-memory
datastore, and wrote some tests to verify that our Storer
implementations adhere to the semantics of the interface, not just the
type information.