Add v1 of the API and an accountsd daemon.
Create an apiv1 package, which runs the first iteration of the API. At
the moment, it simply exposes the functionality of the Storer over HTTP
with a few niceties built-in, but nothing you wouldn't expect.

Create an accountsd package that will build the binary that actually
runs a server. It basically just bootstraps all our requirements and
pulls together all the packages we already have, tying everything
together and kickstarting the entire process.

Create a build.sh script that will build the accountsd binary, building
in all our version information.

Vendor all the new dependencies for the two new packages.