Add Google Cloud Datastore as a Storer.
Add Google Cloud Datastore as an optional Storer. The code is designed
so the only part susceptible to eventual consistency is the listing of
accounts associated with a profile, and even then, either an account
will be missing, or it will have the latest data, so that's not awful.
This is because Datastore doesn't guarantee immediate consistency of
queries without an Ancestor, and we opted not to use an Ancestor because
limiting our writes to 1/sec wasn't really a thing we wanted to do, and
because it would make it harder to verify the uniqueness of an account
ID. So instead, we retrieve things by key wherever possible, because
that's guaranteed to be immediately consistent. Queries don't return
values, they return keys, which are then retrieved by key. So while the
index is eventually consistent, and some keys may be missing from the
results, the values will always be consistent.