Ensure profile ID uniqueness when registering.
When a user signs up for their very first account, that is an
unauthenticated action; anyone can do it. We'll generate a ProfileID and
associate it with their account, at which point they're registered, and
all future actions will be authenticated against that ProfileID. But
when making that registration, we want to be sure we're not duplicating
a ProfileID already in use. This commit adds code to ensure that, by
adding a IsRegistration property to Account, which should be set to true
when the account is being created as part of a registration.

Users can also add accounts to their existing profile, however, and for
those accounts we want to be able to duplicate the ProfileID. We'll be
authenticating that the user has rights to add accounts to that
ProfileID, unlike when registering. In this case, the IsRegistration
field should be set to false, which will cause no uniqueness check to be
performed.

In PostgreSQL, this is implemented as a multicolumn unique constraint,
and we use True for accounts that are registrations, and NULL for
accounts that are not. Because NULL is not considered equal to other
NULL values, the unique constraint will only trigger on accounts where
IsRegistration is True.

For the memory implementation, we just check whether IsRegistration is
true before checking if the ProfileID is unique or not.