Add auth to handlers.
Add a helper to the apiv1 package to pull the session token from a
header, ensure it's valid, and return either an apiv1.Response or a
session.AccessToken, depending on whether there's a valid token or not.
If there's a response, it should be returned. To facilitate this, I had
to add a Status field to the Response type, allowing me to communicate
whether it should be an http.StatusForbidden or an
http.StatusUnauthorized.

This relies on commit 25c676598a11c63ac70e2008faa3419c8f36d83d from the
sessions package.

Use this helper to control access to our handlers, making sure we're
properly authenticated for all of them.