Fix bug around unauthenticated users.
We were previously checking for an empty session iff there was an error,
which doesn't make sense. The fix was to move the empty session handlng
out of the helper, so we can detect whether a user is logged in or not
without needing to parse the error message returned.