Refactor.
Get rid of the storers package. Move the storers_test.go to the main
package, and make it import the tests it's going to run. Each storer now
gets its own package, and the types in it get renamed to reduce
repetition. Move the storer definition out into its own file. Migrations
and SQL are both stored in the postgres storer directory, as they belong
to that storer. The postgres package now also houses the go generate
comment it needs to generate the migrations package.